New table to store extra tags/keys
We get the OSM tags/keys for matching from the item types in Wikidata.

Sometimes we want to check extra tags/keys without adding them to
Wikidata.

This new table is a place to store them, instead of having them in the
source code.

In the future normal users will be able to add or remove extra
tags/keys.